Spun attachments into seperate go file
